renamed note coord package to notecontext
