+++ src/main/java/notecontext/KeySignature.java
@@ -1,4 +1,4 @@
-package note_coordinate;
+package notecontext;
 
 public class KeySignature {
