+++ src/main/java/notecontext/MidiNote.java
@@ -1,4 +1,4 @@
-package note_coordinate;
+package notecontext;
 
 public class MidiNote {
