+++ src/main/java/notecontext/NamedNote.java
@@ -1,4 +1,4 @@
-package note_coordinate;
+package notecontext;
 
 public class NamedNote {
